Adjusting old bindings, adding a new one and various other changes

+++ src/bindings/perf_event.rs
@@ -2,12 +2,14 @@
 #![allow(non_upper_case_globals)]
 #![allow(non_snake_case)]
 
+use libc::c_ulong;
+
 include!(concat!(
     env!("OUT_DIR"),
     "/perf_event_bindings_generated.rs"
 ));
 
-pub const PERF_EVENT_IOC_DISABLE: u64 = PERF_EVENT_IOC_DISABLE_;
-pub const PERF_EVENT_IOC_ENABLE: u64 = PERF_EVENT_IOC_ENABLE_;
-pub const PERF_EVENT_IOC_PERIOD: u64 = PERF_EVENT_IOC_PERIOD_;
-pub const PERF_EVENT_IOC_RESET: u64 = PERF_EVENT_IOC_RESET_;
+pub const PERF_EVENT_IOC_DISABLE: c_ulong = PERF_EVENT_IOC_DISABLE_ as c_ulong;
+pub const PERF_EVENT_IOC_ENABLE: c_ulong = PERF_EVENT_IOC_ENABLE_ as c_ulong;
+pub const PERF_EVENT_IOC_PERIOD: c_ulong = PERF_EVENT_IOC_PERIOD_ as c_ulong;
+pub const PERF_EVENT_IOC_RESET: c_ulong = PERF_EVENT_IOC_RESET_ as c_ulong;
